requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-text-splitters
openai
python-dotenv
pypdf
faiss-cpu
pydantic
python-multipart